refactor(api): split evaluation routes and shrink oversized api functions
