+++ src/quodeq/api/_evaluation_routes.py
@@ -1,308 +1,21 @@
-"""Evaluation listing, creation, status, and cancellation route registrations."""
+"""Evaluation listing, creation, status, and cancellation route registrations.
+
+Thin facade: the endpoints live in routes_evaluations_list.py and
+routes_evaluations_item.py, and the shared already-scored registry lives in
+_scored_jobs_registry.py. Kept here so existing imports/patches
+(``quodeq.api._evaluation_routes.<name>``) continue to resolve.
+"""
 from __future__ import annotations
 
-import logging
-import threading
-from collections import OrderedDict
-from http import HTTPStatus
-from pathlib import Path
-from typing import Any
+from quodeq.services.score_run import score_completed_evidence  # noqa: F401 — re-export/patch target
 
-from flask import Flask, Response, jsonify, request
-
-from quodeq.api._evaluation_helpers import (
-    _build_evaluation_options,
-    _check_eval_rate_limit,
-    _sanitize_url,
-    _validate_ai_cmd,
-    _validate_ai_cmd_path,
-    _validate_ai_model,
+from quodeq.api._scored_jobs_registry import (  # noqa: F401
+    _scored_jobs,
+    _scored_jobs_lock,
+    _SCORED_JOBS_MAX,
+    _claim_scoring,
+    _release_scoring,
+    reset_scored_jobs,
 )
-from quodeq.api.helpers import error_response, scan_target_error, validate_evaluation_payload
-from quodeq.shared.serialization import to_camel_dict
-from quodeq.assistant import get_provider_configs
-from quodeq.api.routes import _reports_dir
-from quodeq.services.active_evaluation import find_active_evaluation
-from quodeq.services.background import BackgroundRunner, ThreadBackgroundRunner
-from quodeq.services.base import ActionProvider
-from quodeq.services.score_run import score_completed_evidence
-from quodeq.services.scan_progress import build_scan_progress
-from quodeq.services.run_events import read_run_dim_states
-from quodeq.shared.utils import is_repo_url
-
-_logger = logging.getLogger(__name__)
-
-# ---------------------------------------------------------------------------
-# Atomic, bounded "already-scored" registry
-# ---------------------------------------------------------------------------
-# Keeps track of job_ids whose background scoring has already been claimed so
-# that repeated GETs for the same job never spawn more than one scoring thread.
-#
-# Uses an OrderedDict as a bounded LRU so the registry cannot grow without
-# limit on a long-running server.  Access is serialised by a Lock so the
-# check-then-add is atomic (closing the TOCTOU race that a plain ``set``
-# would have).
-_scored_jobs: "OrderedDict[str, None]" = OrderedDict()
-_scored_jobs_lock = threading.Lock()
-_SCORED_JOBS_MAX = 1000
-
-
-def _claim_scoring(job_id: str) -> bool:
-    """Atomically claim *job_id* for one-time background scoring.
-
-    Returns ``True`` if this caller should start the scoring thread;
-    ``False`` if another caller already claimed it.
-
-    The registry is bounded to ``_SCORED_JOBS_MAX`` entries (LRU eviction)
-    so memory usage stays constant regardless of server uptime.
-    """
-    with _scored_jobs_lock:
-        if job_id in _scored_jobs:
-            return False
-        _scored_jobs[job_id] = None
-        while len(_scored_jobs) > _SCORED_JOBS_MAX:
-            _scored_jobs.popitem(last=False)  # evict oldest entry
-        return True
-
-
-def _release_scoring(job_id: str) -> None:
-    """Release a claim taken by :func:`_claim_scoring`.
-
-    Used when a discard-cancel claims the slot up front but the cancel
-    itself fails: without the release, a later legitimate cancel of the
-    same job would never get its completed dims scored.
-    """
-    with _scored_jobs_lock:
-        _scored_jobs.pop(job_id, None)
-
-
-def _background(app: Flask) -> BackgroundRunner:
-    """The app's background-task runner. ``create_app`` instantiates it;
-    setdefault keeps bare test apps (register_evaluation_item_routes on a
-    plain Flask) working."""
-    return app.extensions.setdefault("background", ThreadBackgroundRunner())
-
-
-def _read_dim_states(job: Any) -> dict[str, dict[str, Any]]:
-    """Read dimensions.json for *job*'s run dir, returning the dimensions map.
-
-    Empty dict on missing/corrupt file (read_dimensions handles that).
-    """
-    project = getattr(job, "output_project", None)
-    run_id = getattr(job, "output_run_id", None)
-    if not project or not run_id:
-        return {}
-    return read_run_dim_states(_reports_dir(), project, run_id)
-
-# Cap on /api/evaluations ?limit= so a client cannot ask the server to materialize
-# an unbounded list. limit=0 still means "no client cap" but we clamp the actual
-# value the provider sees. 1000 is well above any realistic dashboard query.
-_EVALUATIONS_LIST_HARD_CAP = 1000
-
-
-def register_evaluation_list_routes(app: Flask, provider: ActionProvider, eval_rate_store: object | None = None) -> None:
-    """Register evaluation listing and creation routes."""
-
-    @app.get("/api/evaluations")
-    def list_evaluations() -> Response:
-        raw_limit = request.args.get("limit", 0, type=int)
-        if raw_limit <= 0 or raw_limit > _EVALUATIONS_LIST_HARD_CAP:
-            limit = _EVALUATIONS_LIST_HARD_CAP
-        else:
-            limit = raw_limit
-        state_arg = request.args.get("state", "").strip()
-        states = {s for s in (v.strip() for v in state_arg.split(",")) if s} or None
-        items = provider.list_evaluations(limit=limit, reports_dir=_reports_dir(), states=states)
-        return jsonify([to_camel_dict(j) for j in items])
-
-    @app.get("/api/evaluations/active")
-    def get_active_evaluation() -> Response:
-        """Return the first non-stale running evaluation job, or JSON null.
-
-        Single authoritative answer to "is an evaluation actually running":
-        the staleness rule lives in services.active_evaluation, so shells
-        (native window, frontend) consume it instead of re-deriving it.
-        """
-        job = find_active_evaluation(provider, _reports_dir())
-        return jsonify(to_camel_dict(job) if job is not None else None)
-
-    @app.post("/api/evaluations")
-    def start_evaluation() -> Response | tuple[Response, int]:
-        rate_error = _check_eval_rate_limit(eval_rate_store)
-        if rate_error is not None:
-            return rate_error
-        payload = request.get_json(silent=True) or {}
-        validation_error = validate_evaluation_payload(payload)
-        if validation_error:
-            body, status = error_response(validation_error, HTTPStatus.BAD_REQUEST, "INVALID_INPUT")
-            return jsonify(body), status
-        ai_cmd = payload.get("aiCmd") or None
-        ai_cmd_error = _validate_ai_cmd(ai_cmd)
-        if ai_cmd_error is not None:
-            return ai_cmd_error
-        ai_cmd_path_error = _validate_ai_cmd_path(ai_cmd, payload.get("aiCmdPath") or None)
-        if ai_cmd_path_error is not None:
-            return ai_cmd_path_error
-        model_error = _validate_ai_model(
-            ai_cmd, payload.get("aiModel") or None, get_provider_configs(),
-        )
-        if model_error is not None:
-            return model_error
-        repo = payload.get("repo")
-        _logger.info("start_evaluation: repo=%s, remote_addr=%s", _sanitize_url(repo), request.remote_addr)
-        try:
-            options = _build_evaluation_options(payload)
-        except ValueError as exc:
-            body, status = error_response(str(exc), HTTPStatus.BAD_REQUEST, "INVALID_INPUT")
-            return jsonify(body), status
-        # Same allowlist as /api/scan and POST /api/projects: starting an
-        # evaluation registers + scans the directory and persists its file
-        # tree, so an unvalidated local path would leak arbitrary readable
-        # directories through project endpoints.
-        try:
-            is_url = is_repo_url(str(repo))
-        except ValueError:
-            body, status = error_response("Invalid repo URL", HTTPStatus.BAD_REQUEST, "INVALID_REPO_URL")
-            return jsonify(body), status
-        if not is_url:
-            err = scan_target_error(str(repo), _reports_dir())
-            if err is not None:
-                body, status = err
-                return jsonify(body), status
-        try:
-            job = provider.start_evaluation(repo=repo, reports_dir=_reports_dir(), options=options)
-        except (FileNotFoundError, ValueError):
-            body, status = error_response(
-                "Invalid repository. Provide a local path or a URL like https://github.com/owner/repo.",
-                HTTPStatus.BAD_REQUEST, "INVALID_INPUT",
-            )
-            return jsonify(body), status
-        return jsonify(to_camel_dict(job)), HTTPStatus.ACCEPTED
-
-
-def register_evaluation_item_routes(app: Flask, provider: ActionProvider) -> None:
-    """Register single-evaluation status and cancel routes."""
-
-    def reset_scored_jobs() -> None:
-        """Clear the scored-jobs registry. Useful for test isolation."""
-        with _scored_jobs_lock:
-            _scored_jobs.clear()
-
-    app.extensions["reset_scored_jobs"] = reset_scored_jobs
-
-    @app.get("/api/evaluations/<job_id>")
-    def get_evaluation(job_id: str) -> Response | tuple[Response, int]:
-        job = provider.get_evaluation_status(job_id, reports_dir=_reports_dir())
-        if not job:
-            body, status = error_response("Job not found", HTTPStatus.NOT_FOUND, "NOT_FOUND")
-            return jsonify(body), status
-        # Score any completed dimensions from failed/cancelled jobs (once).
-        # Offloaded to a background thread so the GET returns immediately;
-        # scoring may involve heavy I/O (reading evidence, writing score files).
-        # _claim_scoring() is atomic: exactly one concurrent GET wins the claim.
-        job_status = getattr(job, "status", None)
-        if job_status in ("failed", "cancelled") and _claim_scoring(job_id):
-            _reports = _reports_dir()
-            _score_args = {
-                "outputProject": job.output_project,
-                "outputRunId": job.output_run_id,
-            }
-
-            def _score_in_bg() -> None:
-                try:
-                    score_completed_evidence(_reports, _score_args)
-                except Exception as exc:
-                    _logger.debug(
-                        "Could not score cancelled dimension for %s: %s",
-                        _score_args.get("outputRunId"), exc,
-                    )
-
-            _background(app).submit(_score_in_bg, name=f"score-{job_id}")
-        payload = to_camel_dict(job)
-        payload["dimStates"] = _read_dim_states(job)
-        return jsonify(payload)
-
-    @app.get("/api/evaluations/<job_id>/progress")
-    def get_evaluation_progress(job_id: str) -> Response | tuple[Response, int]:
-        """Return live progress for a scan (works for internal and external runs)."""
-        run_dir = provider.get_log_run_dir(job_id) if hasattr(provider, "get_log_run_dir") else None
-        if run_dir is None:
-            body, status = error_response("Job not found", HTTPStatus.NOT_FOUND, "NOT_FOUND")
-            return jsonify(body), status
-        # Total time limit for the whole run. The snapshot carries the
-        # budget for both internal jobs (JobManager) and index-served runs
-        # (read from status.json). 0 = unlimited -> no budget shown.
-        time_limit_s: int | None = None
-        snapshot = provider.get_evaluation_status(job_id, reports_dir=_reports_dir())
-        if snapshot is not None:
-            raw = getattr(snapshot, "time_limit_s", None)
-            if isinstance(raw, int) and raw > 0:
-                time_limit_s = raw
-        progress = build_scan_progress(
-            job_id, run_dir, time_limit_s=time_limit_s,
-            compiled_dir=Path(app.config["STANDARDS_COMPILED_DIR"]),
-        )
-        if progress is None:
-            body, status = error_response("Run not ready", HTTPStatus.NOT_FOUND, "NOT_FOUND")
-            return jsonify(body), status
-        return jsonify(to_camel_dict(progress))
-
-    @app.delete("/api/evaluations/<job_id>")
-    def cancel_or_delete_evaluation(job_id: str) -> Response | tuple[Response, int]:
-        """DELETE on a running job cancels it. DELETE on a finished job removes it from history.
-
-        Query: ``?intent=cancel|delete`` declares what the client is asking
-        for. Without it, the action is inferred from the momentary status
-        (legacy behavior), which is race-prone: a run finishing while the
-        cancel dialog is open, or a double-clicked cancel, used to fall
-        through to the permanent-purge branch and erase a run the user
-        chose to keep. With intent=cancel this endpoint can never purge;
-        with intent=delete it never silently cancels.
-
-        ``?discard=true`` on a cancel also wipes the run entirely so the
-        next run treats the work as never-happened.
-        """
-        snapshot = provider.get_evaluation_status(job_id, reports_dir=_reports_dir())
-        if snapshot is None:
-            body, status = error_response("Job not found", HTTPStatus.NOT_FOUND, "NOT_FOUND")
-            return jsonify(body), status
-        intent = request.args.get("intent", "").lower() or None
-        if intent == "cancel" and snapshot.status != "running":
-            body, status = error_response(
-                "Evaluation already finished. Nothing was cancelled.",
-                HTTPStatus.CONFLICT, "ALREADY_FINISHED",
-            )
-            return jsonify(body), status
-        if intent == "delete" and snapshot.status == "running":
-            body, status = error_response(
-                "Evaluation is still running. Cancel it before deleting.",
-                HTTPStatus.CONFLICT, "STILL_RUNNING",
-            )
-            return jsonify(body), status
-        if snapshot.status == "running":
-            discard = request.args.get("discard", "").lower() == "true"
-            _logger.info(
-                "cancel_evaluation: job_id=%s, discard=%s, remote_addr=%s",
-                job_id, discard, request.remote_addr,
-            )
-            if discard:
-                # Claim the one-time scoring slot BEFORE the job flips to
-                # cancelled: otherwise the UI's next status poll sees the
-                # cancelled state and spawns _score_completed_evidence,
-                # resurrecting a run the user just discarded.
-                _claim_scoring(job_id)
-            ok = provider.cancel_evaluation(
-                job_id, reports_dir=_reports_dir(), discard_partial=discard,
-            )
-            if not ok:
-                if discard:
-                    _release_scoring(job_id)
-                body, status = error_response("Could not cancel job", HTTPStatus.CONFLICT, "CONFLICT")
-                return jsonify(body), status
-            return jsonify({"ok": True, "action": "cancelled", "discarded": discard})
-        _logger.info("delete_evaluation: job_id=%s, remote_addr=%s", job_id, request.remote_addr)
-        ok = provider.delete_evaluation(job_id, reports_dir=_reports_dir())
-        if not ok:
-            body, status = error_response("Job could not be deleted", HTTPStatus.NOT_FOUND, "NOT_FOUND")
-            return jsonify(body), status
-        return jsonify({"ok": True, "action": "deleted"})
+from quodeq.api.routes_evaluations_list import register_evaluation_list_routes  # noqa: F401
+from quodeq.api.routes_evaluations_item import register_evaluation_item_routes  # noqa: F401
