+++ tools/size_baseline.txt
@@ -8,24 +8,19 @@ src/quodeq/_cli_evaluation.py:467:function
 src/quodeq/_cli_evaluation.py:574:function
 src/quodeq/_cli_resolution.py:1:file
 src/quodeq/_cli_resolution.py:253:function
-src/quodeq/api/_evaluation_routes.py:103:function
-src/quodeq/api/_evaluation_routes.py:130:function
-src/quodeq/api/_evaluation_routes.py:183:function
-src/quodeq/api/_evaluation_routes.py:1:file
-src/quodeq/api/_evaluation_routes.py:251:function
 src/quodeq/api/_import_validation.py:88:function
 src/quodeq/api/_log_stream_routes.py:110:function
 src/quodeq/api/_log_stream_routes.py:135:function
 src/quodeq/api/_run_event_stream.py:1:file
 src/quodeq/api/_run_event_stream.py:289:function
-src/quodeq/api/_sse_log_helpers.py:42:function
-src/quodeq/api/app.py:74:function
 src/quodeq/api/assistant_session_routes.py:58:function
 src/quodeq/api/assistant_turn_routes.py:95:function
 src/quodeq/api/assistant_workspace_routes.py:25:function
 src/quodeq/api/llm_bridge_routes.py:54:function
+src/quodeq/api/routes_evaluations_item.py:95:function
+src/quodeq/api/routes_evaluations_list.py:55:function
 src/quodeq/api/routes_findings.py:90:function
-src/quodeq/api/routes_project_data.py:15:function
+src/quodeq/api/routes_project_data.py:24:function
 src/quodeq/api/routes_project_list.py:1:file
 src/quodeq/api/routes_project_list.py:235:function
 src/quodeq/api/routes_project_list.py:66:function
@@ -33,9 +28,7 @@ src/quodeq/api/routes_shared_config.py:27:function
 src/quodeq/api/routes_shared_mirrors.py:41:function
 src/quodeq/api/routes_shared_pull.py:27:function
 src/quodeq/api/routes_shared_pull.py:30:function
-src/quodeq/api/standards_import_routes.py:15:function
 src/quodeq/api/standards_overrides_routes.py:29:function
-src/quodeq/api/standards_visibility_routes.py:30:function
 src/quodeq/api/terminal_routes.py:206:function
 src/quodeq/api/terminal_routes.py:90:function
 src/quodeq/api/zip.py:72:function
